LayoutTests/platform/mac-snowleopard/canvas/philip/tests/2d.pattern.image.undefined-expected.txt

< [index] >
2d.pattern.image.undefined
References: 2d.pattern.wrongtype
Actual output:
Failed assertion: expected exception of type TYPE_MISMATCH_ERR, got: Type error
